Add Variableratio/interval flag in headfix gui
Clean up functional go/nogo task; update parameters in all files

+++ Namlab_behavior_lickforreward/Namlab_behavior_lickforreward.ino
@@ -159,8 +159,8 @@ unsigned long CS_t_fxd[2 * numCS];
 unsigned long CSpulse[numCS];
 unsigned long CSspeaker[numCS];
 unsigned long CSlight[numCS];
-unsigned long golickreq[numCS];
-unsigned long golicktube[numCS];
+signed long golickreq[numCS];
+int golicktube[numCS];
 unsigned long CSsignal[numCS];
 unsigned long meanITI;           // mean duration of ITI for the exponential distribution OR minimum ITI for uniform distribution
 unsigned long maxITI;            // maximum duration of ITI
